p3_air

Trait Air

pub trait Air<AB: AirBuilder>: BaseAir<AB::F> {
    // Required method
    fn eval(&self, builder: &mut AB);
}

An algebraic intermediate representation (AIR) definition.

Contains an evaluation function for computing the constraints of the AIR. This function can be applied to an evaluation trace in which case each constraint will compute a particular value or it can be applied symbolically with each constraint computing a symbolic expression.

Required Methods

fn eval(&self, builder: &mut AB)

Evaluate all AIR constraints using the provided builder.

The builder provides both the trace on which the constraints are evaluated on as well as the method of accumulating the constraint evaluations.

Arguments

builder: Mutable reference to an AirBuilder for defining constraints.