p3_air

Trait PermutationAirBuilder

pub trait PermutationAirBuilder: ExtensionBuilder {
    type MP: Matrix<Self::VarEF>;
    type RandomVar: Into<Self::ExprEF> + Copy;

    // Required methods
    fn permutation(&self) -> Self::MP;
    fn permutation_randomness(&self) -> &[Self::RandomVar];
}

Trait for builders supporting permutation arguments (e.g., for lookup constraints).

Required Associated Types

type MP: Matrix<Self::VarEF>

Matrix type over extension field variables representing a permutation.

type RandomVar: Into<Self::ExprEF> + Copy

Randomness variable type used in permutation commitments.

Required Methods

fn permutation(&self) -> Self::MP

Return the matrix representing permutation registers.

fn permutation_randomness(&self) -> &[Self::RandomVar]

Return the list of randomness values for permutation argument.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.